requirements.txt
dash==3.0.4
plotly==6.1.2
pandas==2.2.3
folium==0.19.7
branca==0.8.1
geopy==2.4.1
python-dotenv==1.1.0
sayari==0.1.41
